spi_qsd: Fix potential race condition
For newer versions of the MSM SPI controller, the outputservice IRQ will
fire on the transition of the operational state from RESET to RUN even if
the output FIFO is empty.  The intention is that the software can write to
the output FIFO from interrupt context after putting the controller into
the RUN state.

This driver needs to continue to support the older SPI controller as well
as the new version of the controller.  Therefore, SPI transactions need
to be started by writing the first word of data to the output FIFO from
task context.  This can result in a race condition for the newer
controller if the outputservice IRQ occurs while data is being read, in
task context, from the input buffer and before it has been written to the
output FIFO.  To eliminate the possible race condition, data needs to be
written to the output FIFO before changing the operational state to RUN.
If the race condition occurs, data can be lost or written out of order.

Signed-off-by: Kenneth Heitke <kheitke@quicinc.com>