diag: Enable user space client kill read thread
Each user space app has a read thread associated with the diag driver to
communicate masks change requests from kernel to user space. When the app
wants to discontinue using DIAG, it calls DIAG's DeInit function which
in turn has to kill the read thread. However, this read thread is in an
infinite loop and has a blocking system call. This change enables the
DeInit function to send an ioctl to kernel, enables the system call to
return and thus break out of the read thread.

Signed-off-by: Shalabh Jain <shalabhj@quicinc.com>
CRs-fixed: 212158