msm: audio: 7x30: send voice responses after devices are opened
During 1x voice call test, the application processor sometimes received
the messages from modem and sent the response to the modem before the
voice devices were opened. At this senario, opening devices later will
get failure. We got AFE enable failure.  This patch is to check if the
voice devices are opened before sending responses to modem, if not, wait
for voice devices to be opened within 30 seconds.

Signed-off-by: Helen Zeng <xiaoyunz@quicinc.com>