usb: gadget: add PMIC Notification support for OTG
7x27 and 8650 FFA HW do not have external LDO to power USB PHY3.3V. The
reason for this is, extra component on the board and require to turn ON
PHY comparators always, which leads more PHY power consumption. PMIC has
capability to detect VBUS SN changes and can notify Apps. To avoid
external LDO component and to reduce PHY power consumption, turn off PHY
comparators as a part of cable disconnect and depend on PMIC for VBUS SN
notifications.

This solution works only for peripheral only mode and doesn't work for
OTG mode, because ID line is not routed to PMIC and hence PMIC cannot
detect ID change. Hence, in OTG mode, require external LDO and turn ON
comparators always.

CRs-fixed: 213227

Signed-off-by: Pavankumar Kondeti <pkondeti@qualcomm.com>