Marimba: Regional settings support for FM core
Regional settings are used to configure FM core based on the
geographical location. The FM hardware supports two native
frequency bands (US and Japan) but can be configured to any custom
band in the range of 76-108 MHz (for e.g. Japan wide band). Channel
spacing and emphasis also needs to be configured as part of
regional settings.

Signed-off-by: Mangesh Pardeshi <mangeshp@quicinc.com>