mfd: pmic8058: add set_wake and suspend/resume support
The set_wake method is to provide the support for drivers to set or
unset wake-up capability for their interrupts. The suspend and resume
methods as I2C driver are to disable/enable corresponding interrupts
in PMIC SSBI registers. If there is no wakeup interrupt, whole PMIC
interrupt is disabled.

Signed-off-by: Willie Ruan <wruan@quicinc.com>