msm: board-7x30: Added support for SD/MMC on FLUID
Slot 1 on FLUID needs to support 2.85 volts operation and we have a level
shifter to drive the VDD on this slot at that voltage. This level shifter
takes input from LDO5 and is enabled via GPIO 35. So both LDO5 and GPIO35
must be enabled for using slot 1 on FLUID.

In addition, for FLUID the voltage capability for slot-1 is set to 2.85V.

Signed-off-by: Murali Palnati <palnatim@qualcomm.com>