usb: gadget: Enable usbh_clk only during core and phy reset
usbh_clk is not required for normal usb operation and its only
required while resetting usb core and phy; By keeping it on
when usb is active, ~30ma of higher hsdpa current is observed.
As this clock is not required for normal usb operation and to
reduce power consumption during hsdpa enable usbh_clk only during
core and phy reset.

Signed-off-by: Velempati Chiranjeevi <c_cvelem@qualcomm.com>