msm: dal: allow async support for dalrpc_fcn_5
Currently dalrpc only supports synchronous function
calls. Support is being added to allow asynchronous
function calls be made for dalrpc_fcn_5.

To use this function as async, bit 31 for parameter
ddi_idx needs to be set to 1. This change will make
the dalrpc_fcn_5 as non-blocking. If bit 31 for
ddi_idx is set as 0 it will remain as a blocking call.

Signed-off-by: Suresh Vankadara <svankada@qualcomm.com>