usb: msm_gadget: Register with switch to send online/offline uevents
USB online/offline uevents are currently sent by mass storage function
driver.  Userspace applications depend on this uevent to know the
usb state.  This does not work if mass storage function is not present
in the usb composition.  Hence send uevents based on the connection
status (pullup enabled/disabled) of usb.  When there is a composition
change, an offline uevent followed by an online uevent will be sent.

Signed-off-by: Pavankumar Kondeti <pkondeti@qualcomm.com>