Define and process new type of memory tag (ATAG_MEM_RESERVED)
This memory tag represents memory which can be powered on and off.
During kernel initialization, it is not added to the kernel page pool,
but must be memory hotplugged by userspace. The kernel does need to
know of the existence of this memory so that it is mapped properly into
the kernel's virtual address space, which must be done during
initialization so that the needed virtual addresses aren't used for
other purposes.

Change-Id: Iece13619f3226bdb9b83a8376b33199b898f6d7d
Signed-off-by: Larry Bassel <lbassel@quicinc.com>