diag: Add spin lock around critical section in timer routine
DIAG char driver uses a timer to drain data to USB. This timer accesses
a critical section also accessed by driver API. Until now, the driver
API just disabled timer before accessing critical section and accquired
the spin lock. The timer routine was missing the spinlock. Hence, adding
the spinlock in timer routine as well. Here, the timer tries to get the
spinlock. If it doesnt get it, it means, the driver API is executing the
critical section. Hence, the timer need not do anything and it exits.

Signed-off-by: Shalabh Jain <shalabhj@quicinc.com>