msm: audio: 7x30: Enable headset power amplifier
Problem:
When Marimba CODEC is configured to headset mode, ncp
voltage regulator is not enabled to power up headset
amplifier. No audio would be heard without ncp on if
target is capless mode

Change:
Update Marimba based headset devices to accommodate
the requirement to turn on NCP VREG when Marimba CODEC
is configured to headset mode. This only applies to
capless devices such as the FFA, and not the SURF.

CRs-fixed: 217529 217792

Signed-off-by: Patrick Lai <plai@quicinc.com>