msm: board-7x30: Advertise 2.85 Volts for SDIO slot 3
SDIO slot 3 on 7x30 is used by the WLAN module which
advertises 2.85v operation capability. The SDIO slot on
the MSM platform should therefore also advertise the same
voltage (2.85v) or else the SD Bus driver will reject
the WLAN card because of voltage mismatch between what is
available on the slot and what the card supports.

Change-Id: Ice79316e1fca6029bd077d9593243e30d8d98df3
Acked-by: Kumar Anand <kanand@qualcomm.com>
Acked-by: Yunsen Wang <yunsenw@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>