pm_qos: Support alternate class update functions.
The ability to call custom update functions, in place of the default
pm_qos update function, can be useful when pm_qos's voting and
comparitor features are not required.  This allows the pm_qos APIs
to be used as wrappers for arbitrary resource management code.

Custom add, remove and update functions can be created, conforming
to the prototype:
  int my_func(struct pm_qos_object *class, char *class_name,
  		s32 value, void **req_data)

A void pointer is added to 'struct requirement_list' for use by update
functions which may require the ability to share requirement-specific
data.  A pointer to this is available as an argument to the update
functions.

Custom update functions may be assigned to a class by setting the
'add_fn', 'update_fn' and 'remove_fn' function pointers for each
class.  These are called when pm_qos_add_requirement,
pm_qos_update_requirement, and pm_qos_remove_requirement are,
respectively, called.  If a custom function is not assigned the
default update function is used.

Change-Id: I28fe4ce64690c17a7919cf47635e83ef07e12f08
Signed-off-by: Matt Wagantall <mattw@quicinc.com>