usb: gadget: Implement init and exit routines in ADB
When ADB daemon runs it expects adb device node to be present
even when ADB is not in composition.  ADB daemon will be blocked
till ADB interface is available.  Move misc driver registration
to an init routine and make ADB device node available even when
ADB interface is removed from composition. This patch also ensures
that cdev is initalized properly.

Signed-off-by: Pavankumar Kondeti <pkondeti@qualcomm.com>