diag: Add a printk to warn while dropping packet
DIAG char driver receives data from USB and verifies that it is HDLC
encoded correctly before forwarding it to modem. If the HDLC encoding is
not correct, the packet is dropped. This print message will show up when
the packet is dropped and will help debugging as to why a packet did not
reach modem.

Signed-off-by: Shalabh Jain <shalabhj@quicinc.com>