msm: Change type of msm_shared_ram_phys to 'unsigned int'.
The __phys_to_pfn(msm_shared_ram_phys) macro is called, which performs
a right-shift of its argument. If the argument is of a signed type,
then its MSB is used for sign extension, corrupting the address if
msm_shared_ram_phys's MSB is set to '1'.

Hence, change the type to 'unsigned int' to support addresses which
have their MSB set.

Change-Id: I8282671ae8353f87275313424e347e79f347b208
Signed-off-by: Matt Wagantall <mattw@quicinc.com>