diag: Edit SMD callback functions
DIAG char driver uses SMD channels to pull in data from modem and Q6. It
invokes SMD functions when it is ready to receive more data. These
functions can be called either when SMD notifies it has some data OR
when USB is done writing data. Hence these functions can be called
either in SMD's callback context or in any other context. There are
different functions for each of the above mentioned context. Earlier,
DIAG was calling same function in both contexts which could lead to an
error.

Signed-off-by: Shalabh Jain <shalabhj@quicinc.com>