diag: Release mutex if client limit is reached
DIAG char driver maintains a table to keep track of all clients using
DIAG services. This table is protected by a mutex while it is modified.
Fixed a corner case, where if the max client limit is reached, the mutex
is released. Without this change, if the limit is reached, the mutex
remains locked.

Signed-off-by: Shalabh Jain <shalabhj@quicinc.com>
CRs-fixed: 220788