msm: kgsl: Add barriers and flush the tlb when changing MMU mappings
Memory barriers are required whenever we map/unmap entries from
pagetables because the pagetables are allocated from normal
noncached memory pool.

The tlb of the GPU needs to be flushed when entries from its
pagetable are unmapped as we do not want old/invalid entries to remain
in the tlb.

Signed-off-by: Shubhraprakash Das <sadas@quicinc.com>