PM: Add NPA Remote support
Provides support to make requests to NPA resources on another
processor domain. A 'remote' resource is a NPA resource in the other
processor. A local resource is a mirror of the remote resource in the
current processor. A local resource can do the aggregation of requests
or can relay the request directly to the remote resource.

A local resource of a 'remoted' resource, once created, acts like any
other regular NPA resource in the same domain and and allows clients
and events to be created and registered for.

Creation of a local resource is achieved by declaring the resource
as either a local aggregation resource or remote aggregation and
pass the so defined data structures to npa_remote_define_node().

Signed-off-by: Praveen Chidambaram <pchidamb@quicinc.com>