msm: add msm specific support for memory on and low-power states
This is done by calling into a new function which will make an
NPA call into the modem's NPA memory node.

Also, the existing power off code has been refactored to call
the same function.

This function is currently a stub which writes a bit
pattern over "powered off" memory and does nothing in
the power on and low power cases.

Change-Id: I91e3c4146258520f7a995ee028ee3ac00beef8df
Signed-off-by: Larry Bassel <lbassel@quicinc.com>