msm: board-msm7x30: Make audio definitions dependent on MSM7KV2_AUDIO.
A number of functions and structures are only used if MSM7KV2_AUDIO,
so they should not be built into the kernel when MSM7KV2_AUDIO is
disabled. The msm_snddev_* were incorrectly placed inside an ifdef
MSM_CAMERA. Also, msm_snddev_init is not defined (and thus should not
be called) when MSM7KV2_AUDIO is disabled.

Change-Id: I260792d6889e46a70cc646151ee8cc5b83a71870
Signed-off-by: Isaac Sutherland <isaacs@quicinc.com>