diag: Increase buffer size for write requests from clients
DIAG char driver uses a mempool to copy write requests from the clients.
The item size for this mempool was not large enough to accommodate big
size packets. Increasing the item size to a reasonable number and
decreasing the total number of items, so that we don't use too much
memory.

Signed-off-by: Shalabh Jain <shalabhj@quicinc.com>