msm: audio: 7x30: Prevent HSED_MICBIAS getting disabled
Problem:
Headset switch detection does not work after 1st try.
The reason is that HSED_MICBIAS is getting disabled when
headset TX device is disabled. 7x30 relies on HSED_MICBIAS
to enabled in order to detect headset plug-in.

Change:
Remove micbias controller 1 out of list of micbias to be
enabled for headset_mono_tx and tty_headset_mono_tx

CRs-Fixed: 222110, 223630
Signed-off-by: Patrick Lai <plai@quicinc.com>