msm: audio: 7x30: Add wakelock mechanism in audpp and audpreproc services
wakelock mechanism is handled in adsp driver. If all modules
are in the disabled state, it releases wakelock to allow apps
processor to enter into suspend power collapse state. AFETASK
is in the enabled state for voice call, but no activity from Apps
processor. So we can allow apps processor to enter into suspend
state, but ADSP driver doesn't have this information. It prevents
the adsp driver to release the wakelock.

Move this logic to audpp and audpreproc services. They know  the state
of audio sessions. Based on this information they hold and release locks.

CRs-Fixed: 224878

Signed-off-by: Suresh Vankadara <svankada@qualcomm.com>