msm: adsp: 7x30: Remove wakelock mechanism in adsp driver
wakelock mechanism is handled in adsp driver. If all modules
are in the disabled state, it releases wakelock to allow apps
processor to enter into suspend power collapse state. AFETASK
is in the enabled state for voice call, but no activity from Apps
processor. So we can allow apps processor to enter into suspend
state, but ADSP driver doesn't have this information. It prevents
the adsp driver to release the wakelock.

Remove this logic and will be added in audio service.

Signed-off-by: Suresh Vankadara <svankada@qualcomm.com>