camera: Fix for RAW snapshot crash on 8K.
During RAW snapshot sometimes the CAMIF doesn't stop after
the snapshot is done there by sending more frames to the VFE.
This causes the VFE to overwrite the snapshot buffer again. If the
timing of the VFE to overwrite the extra snapshot data to the buffer
coincides with the memcopy operation of the snapshot buffer by the
application it will cause camera to crash. This is fixed by sending
a HALT command to the CAMIF hardware when the snapshot is done.

CRs-Fixed: 222064

Signed-off-by: Ninad Patgaonkar <ninadp@quicinc.com>