msm: Fix for camframe timeout issue.
Before the start recording command was executed completely
the reg update irq happened and due to this the pong address
was not updated properly and that leads kernel to behave abnormally.
So the approach for this fix is to update the ping and pong when reg
update irq happens and also when the recording is enable.

CRs-fixed: 225374

Signed-off-by: Azam Kapatrala Syed <c_azams@quicinc.com>