camera: Fix for 8K target freeze on camera/camcorder exit.
The work queue was accessing NULL pointer on camera exit.
This was happening because the camera driver upon exit frees
all the memory and sets the pointers to NULL. Since the work
queue has lower priority it was getting executed after the
camera driver exits there by accessing NULL pointer. Added a check
in the work queue so that the work queue doesn't execute after
camera driver exits.

CRs-Fixed: 225616

Signed-off-by: Ninad Patgaonkar <ninadp@quicinc.com>