camera: Fix for NULL pointer exception in 8K camera driver.
The IRQ work queue was accessing NULL pointer on camera exit.
This was happening because the camera driver upon exit frees
all the memory and sets the pointers to NULL. Since the work
queue has lower priority it was getting executed after the
camera driver exits there by accessing NULL pointer. Added
spinlock in the work queue to fix this problem.

Signed-off-by: Ninad Patgaonkar <ninadp@quicinc.com>