msm: clock-7x30: Fix PLL status register access
The PLL status register doesn't have hardware copies and hence
cannot be read from more than one base/shadow region at a time.
The register needs to be shared with the other processors, so the
ownership remains with the clock control base region.

Since we need to read the PLL status register, we also need to
map in the clock control base region.

Change-Id: Ia386bc7c110305b0e7ecd20e49618b5c010734c6
Signed-off-by: Stephen Boyd <sboyd@quicinc.com>