mm: arm: Prevent page movement on forbidden pfn ranges
When low on pages, the memory subsystem attempts to find pages in
other movable lists. When doing so, it attempts to find and move
4 MiB blocks of pages. The code in move_freepages_block
presumptuously assumes that there is only one memory bank in each
memory zone, or that each memory partition ends on one less than
a 4 MiB aligned address. This is not the case in all memory maps.

As a temporary solution, we will walk the memory banks provided
by the ARM subsystem to verify we aren't trying to move pages
that overlap with holes in the address map.

Signed-off-by: Michael Bohan <mbohan@quicinc.com>