msm-nand: Add support for interleaving mode
For platforms that contain two nand controllers, better performance
can be achieved by using both the controllers to read and write pages.
For a page read/write, the codewords are read/write to different
nand chips in a interleaved fashion which are hooked to different
chipselects.These code words can be handled by each controller and
some activities can happen in parallel, thereby reducing the overall
time for a page read/write.

Signed-off-by: Murali Nalajala <mnalajal@qualcomm.com>