pmem: Add device names to pmem debug logs
pmem logs only the device ids in the log messages.
The id needs to be matched with the device names on
each target, as the same id can represent different pmem
devices across targets. Logging the device name in the logs
avoids the need to look up the name corresponding to the id
and faciliates easy debugging.

Change-Id: Ie9a063e260a1ceef42e803d9c3a95269664dbd91
Signed-off-by: Naveen Ramaraj <nramaraj@quicinc.com>