arm: Add config option for HOLES_IN_ZONE
This option is for FLATMEM configurations with multiple memory
banks that have end addresses not aligned to MAX_ORDER_NR_PAGES.

To save memory, ARM frees memmap regions for pfns associated with
gaps in the address map. When the VM code migrates pages, it will
assume that the memmap entries are valid.

This option adds extra checking with pfn_valid() to
ensure the memmap entries are valid. It may adversely impact
performance.

Change-Id: I7618f7a725e19964e608385a8566202fa0d5235c
Signed-off-by: Michael Bohan <mbohan@quicinc.com>