msm: clock-7x30: Make GLBL_ROOT_ENA a parent "clock"
The GLBL_ROOT_ENA bit in GLBL_CLK_ENA_REG needs to be set for any
global clock to be on. Previously we would just set this bit on
init but if we are going to support dynamic clocks we should only
set this bit if we have ownership of the global clocks. One way
to do this is to make it a parent clock of the global clocks, so
it's only enabled if a global clock is on.

Change-Id: Ic992a92552bc88a7ee180dfac268448ead22df97
Signed-off-by: Stephen Boyd <sboyd@quicinc.com>