usb: msm_gadget: move the wakelock to peripheral driver
As OTG driver invokes either host or peripheral driver in atomic
context based on the ID pin, OTG driver doesn't need to maintain
wakelock for itself.
Since Host Controller driver is having it's own wakelock, remove
wakelocks in the OTG driver and add in the gadget peripheral driver.

This patch removes debugfs support to release the wakelock during bus
suspend in the OTG driver and adds this support in the peripheral
driver.

Signed-off-by: Anji Jonnala <anjir@qualcomm.com>