qup_i2c: Make driver compatible with new QUP I2C core.
8660 and upcoming chipsets have new QUP I2C core which has 1 interrupt
instead of 3. The interrupt is level triggered.
We rely on boards file and interrupt resource to determine how many
interrupts we have. We rely on QUP registers rather than interrupt
number in ISR to figure out interrupt reasons.
The driver changes are backwards compatible and work on chipsets with
older QUP core. (e.g. 7x30)

Signed-off-by: Sagar Dharia <sdharia@quicinc.com>