msm: rpc_pmapp: Add clock vote API
The clock vote API takes three parameters: a voter id, a clock id,
and a voting type. The voter id uniquely identifies the client
must be exactly four characters. The clock id is an enumeration of
the four different clock buffers. The vote type allows the client to
specify whether the clock should be turned off, turned on, or
switched to pin control mode. Pin control mode takes the highest
priority of all the vote types. Only when all clients have voted for
the clock to be off will the modem disable the clock.

Change-Id: I509ef1124e044d6bb9d13bdc3d3018e56d5adf23
Signed-off-by: Bobby Crabtree <bobbyc@quicinc.com>