pmic8058: Change driver initialization macros
The PWM init macro needs to be changed from module_init to
sybsys_initcall so that it is initialized before the drivers that
depend on it. Since PWM is a subdevice of the PMIC8058 core driver,
the core driver and its dependencies needs to be initialized earlier
as well. The GPIO and MPP subdevices initialization priority are also
updated to match that of the PWM module.

Change-Id: I2511a9e2c7eb0130d7ae42931c8edc388c1351e4
Signed-off-by: Bobby Crabtree <bobbyc@quicinc.com>