msm: Vote for vreg s2 from tsadc module on Marimba B0 chips
VREG_RFA(S2) had to be enabled along with LDO24 to have Marimba switched
ON for A0 version.  This change will turn OFF S2 buck if the revision of
Marimba is B0 for power savings. Clients have to enable and disable this
VREG when required.

Voltage regulator "s2" is voted from core driver in A0 chips due to a
hardware bug. B0 chips have the fix so it is upto the child to vote for
"s2" based on need. Tsadc module needs "s2" to be on to function.

Signed-off-by: Karthik Parsha <kparsha@quicinc.com>