msm: clock-7x30: Dynamic clock ownership support
Determine at boot which clocks are locally controlled and which
are remotely controlled. We can determine which clocks are
locally owned by reading their respective ownership bits.  Upon
determing the ownership, we set the ops pointer to either the
local implementation or the proc_comm implementation depending on
if the clock is locally owned or remotely owned.

The benefit of this is it allows the clock ownership to change
without requiring any recompilation or changes on our part.

Note: If a clock can be locally controlled we use the CLK_7X30 macro
(instead of CLK_PCOM) in the clock tables. This allows easy
identifiability of exclusively proc_comm clocks.

Change-Id: I9539d4555078535193a35be12761bd192550ab67
Signed-off-by: Stephen Boyd <sboyd@quicinc.com>