alsa: msm7x30: New control to provide the count of mixer controls
The new control helps in querying the device count from the list of
controls available. The current list of controls are split into to
categories, Simple controls and device controls. The count helps in
user space library to dynamically read the count of simple controls
and hence makes it transparent of number of controls.

Change-Id: Ie172c4e4460fd1c8616095f7917ab1337dff845a
Signed-off-by: Asish Bhattacharya <asishb@qualcomm.com>