msm_kgsl: Migrate to grp source clock
The grp source clock controls the rates for the slave clocks grp_clk and
imem_clk. It's incorrect to call clk_set_rate() on slave clocks
therefore get the source clock and call set_rate() on that. If
this is a target which doesn't support source clocks, fallback to
a slave clock.

Change-Id: If3b2145e1293dc1a922e74fce5ed0178fe2856ab
Signed-off-by: Stephen Boyd <sboyd@quicinc.com>