msm_fb: Migrate to tv source clock
The tv source clock controls the rates for the slave clocks
tv_dac, tv_enc, and hdmi. It's incorrect to call clk_set_rate()
on slave clocks, therefore get the source clock and call
set_rate() on that. If this is a target which doesn't support
source clocks, fallback to the tv_dac slave clock.

Also, remove the clk_get_rate() calls since they're unnecessary.

Change-Id: Ib1712c51310884615d7ab4f3b8731653d2e68d29
Signed-off-by: Stephen Boyd <sboyd@quicinc.com>