media: camera: changes to improve the stability of VFE3.1 driver
Added read back of VFE command registers for work around potential
VFE hardware bug. For VFE command registers, we found issue with the
written commands not being taken to effect until we do a immediate
read. The cause is unknown, however the work around of reading back
the register works.

Change-Id: I59b3a32d28b6c45892aad182c20a9ac6b210106a
Signed-off-by: Nagesh Subba Reddy <c_nsubba@quicinc.com>