logger: Fix logcat invalid read for high volume logs
When the volume of logs is very high, the read offset
is lapped by the write offset. This is taken care of
in the fix_up_readers function.

If fix_up_readers is called just after the spinlock
is released in logger_read but before do_read_log_to_user
reacquires it, the length passed to do_read_log_to_user
becomes invalid and causes the next read to be invalid.

We now re-check the next entry length if we detect that the
reader offsets have been fixed.

Tested on MSM7x27 FFA.

Change-Id: If6f742e4d13cf70528a084c6633be5de47d82c82
Signed-off-by: Naseer Ahmed <naseer@qualcomm.com>