msm: clock-7x30: Support power rails in auto mode
Power rails have two modes: manual and auto. When a rail is in
manual mode, the driver assumes control of enabling and disabling
the rail. When the rail is in auto mode, the clock which is
closest related to the rail is used to determine if the rail
should be turned on or off. For example, if the clock is on the
rail is turned on as well.

This works well when the clock is proc_comm controlled since the
modem can tell when a clock is on or off. When the clocks are
locally owned, the modem will do nothing for rails in auto mode
since it never receives proc_comm requests for the clocks the
rails are related to.

Therefore call into the power rail code from the local clock code
and enable or disable the rails for the associated clock as long
as the rail is not manually controlled. This simulates auto mode
for locally controlled clocks.

Change-Id: I62061306b93fb2f6b2e77915581887f520a76c8b
Signed-off-by: Stephen Boyd <sboyd@quicinc.com>