msm: audio: 7x30: Initialize audpre wakelock only once
Problem.
The current code initialises the global wakelock every time the session is
attempted to open and does not handle the locks properly in case the
session closes with error.

Impact
The same locks are reinitialized repeatedly and not cleared. This results
in the apps waiting on resource once the suspend happens.

Solution
The audpreproc driver is modified to check the current state of wake lock
and initialize it only if not done already.Also the code is modified to
unlock the wake locks if the session closes with an error.

Change-Id: I621008d4209d0983edcd00faa40a0007f749db97
CRs-fixed: 231064,229055
Signed-off-by: Asish Bhattacharya <asishb@qualcomm.com>
Signed-off-by: Rajesha Kini <rkini@qualcomm.com>