msm: clock-7x30: Make freq tables match proc_comm expectations
clk_set_rate() calls over proc_comm are usually expecting the
rate to be exactly what proc_comm has decided they should be. The
only exceptions to this rule are the mdp_lcdc_pclk_clk, vfe_clk,
cam_m_clk and sdc_clk. The sdc_clk expects the frequency to be
within +/- 1000 Hz of the actual value. The other three clocks
choose the closest rate to the rate requested.

The problem occurs when the rates are slightly different. For
example, consider the mdp_clk. When it's local we require 49 Mhz
but proc_comm requires 49.152 Mhz. This will never work for both
the local case and the proc_comm case. Either we use 49 Mhz and
proc_comm fails, or we use 49.152 Mhz and round the rate up
getting back 52 Mhz which isn't the same.

To make things simpler just make the rates for local clocks the
same as what proc_comm is expecting them to be. This way we don't
have problems where the local rate is less than the rate expected
by proc_comm. Doing this makes it simpler for the driver authors
to migrate to local clocks.

Change-Id: I37417ff5db89653487e161fea73e18fb047856d8
Signed-off-by: Stephen Boyd <sboyd@quicinc.com>
Signed-off-by: Matt Wagantall <mattw@quicinc.com>